feat: move QuicInboundConnection to outbound (#3)
