+++ src/northbound/outbound.rs
@@ -1,11 +1,16 @@
 // Copyright (C) 2026 ReteLabs LLC.
 // Licensed under Apache-2.0 or MIT at your option.
 
-use error_stack::ResultExt;
+use async_trait::async_trait;
+use error_stack::{Report, ResultExt};
+use tokio::io::{AsyncRead, AsyncWrite};
 
 use crate::{
     TcpDirectTargets,
-    core::transport::QuicPublisher,
+    core::transport::{
+        QuicPublisher,
+        endpoint::connection::{Accept, QuicConnection},
+    },
     northbound::outbound::tcp::{TcpDirectHandle, TcpDirectOutbound},
     utils::report::ErrorReport,
 };
@@ -15,6 +20,26 @@ pub mod tcp;
 #[derive(Debug, thiserror::Error)]
 #[error("{0}")]
 pub struct Error(String);
+
+trait QuicStream: AsyncRead + AsyncWrite + Unpin + Send {}
+
+impl<T> QuicStream for T where T: AsyncRead + AsyncWrite + Unpin + Send {}
+
+#[async_trait]
+trait QuicInboundConnection: Send + Sync {
+    async fn accept_stream(&self) -> Result<Box<dyn QuicStream>, Report<Error>>;
+}
+
+#[async_trait]
+impl QuicInboundConnection for QuicConnection {
+    async fn accept_stream(&self) -> Result<Box<dyn QuicStream>, Report<Error>> {
+        let (send, recv) = self
+            .accept_bi()
+            .await
+            .change_context(Error("QUIC stream accept failed".into()))?;
+        Ok(Box::new(tokio::io::join(recv, send)))
+    }
+}
 
 /// Dependencies required to construct an [`OutboundBundle`].
 #[fundle::deps]
@@ -48,18 +73,13 @@ async fn init_tcp_direct(
         return Ok(None);
     }
 
-    let services = deps
-        .tcp_direct_targets
-        .0
-        .keys()
-        .cloned()
-        .collect::<Vec<_>>()
-        .join(", ");
-    let tcp_direct = TcpDirectOutbound::new(deps.tcp_direct_targets.0, deps.quic_publisher)
+    let tcp_direct = TcpDirectOutbound::new(deps.tcp_direct_targets.0.clone(), deps.quic_publisher)
         .await
         .change_context(Error("Failed to create TCP direct outbound".into()))?
         .spawn();
 
-    log::info!("TCP direct outbound serving: {services}");
+    for (service_name, addr) in &deps.tcp_direct_targets.0 {
+        log::info!("TCP direct outbound serving '{service_name}' on '{addr}'");
+    }
     Ok(Some(tcp_direct))
 }
